+++ CHANGELOG.rst
@@ -19,6 +19,7 @@ Next version
 - Added a button for editing the raw HTML. This is sometimes useful.
 - Stopped generating source maps unless in dev mode. I like source maps a lot
   in general, but the files are really big in this case.
+- Added a button to the menu to insert horizontal rules.
 
 
 `0.1`_ (2024-03-11)